Refactor comments system in a single block
