--- packages/block-library/src/blocks/post-comments-form/index.php
@@ -1,1 +1,0 @@
-<?php // Silence is golden
